Separated RepositorySelector from SidePanel and moved it into UI
